pyproject.toml
[tool.poetry]
name = "quickbase-client"
version = "0.5.0"
description = "A Quickbase Python API Client Generator"
authors = ["Tim Kutcher <tim@tkutcher.com>"]
maintainers = ["Tim Kutcher <tim@tkutcher.com>"]
readme = "README.rst"
license = "LICENSE.md"
homepage = "https://github.com/tkutcher/quickbase-client"
repository = "https://github.com/tkutcher/quickbase-client"
documentation = "https://quickbase-client.readthedocs.io/en/latest/?badge=latest"
keywords = ["quickbase", "client", "quickbase-api", "orm", "generator"]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers"
]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/tkutcher/quickbase-client/issues"

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.24.0"
stringcase = "^1.2.0"
attrs = "^20.3.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
pytest-cov = "^2.10.1"
requests-mock = "^1.8.0"
flake8 = "^3.8.4"
flake8-quotes = "^3.2.0"
flake8-import-order = "^0.18.1"
pytest-flake8 = "^1.0.6"
tox = "^3.20.1"
Sphinx = "^4.5.0"
sphinx-rtd-theme = "^0.5.0"
python-dotenv = "^0.15.0"
six = "^1.15.0"
pytest-mock = "^3.6.1"
black = "^22.3.0"

[tool.poetry.scripts]
qbc = 'quickbase_client.tools.qbc:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
